@@ -10,7 +10,6 @@
     "test": "jest",
     "ci": "npm run lint && npm run flow && npm test",
     "build-assets": "./bin/build_assets.sh",
-    "precommit": "lint-staged",
     "docs:build": "cd website && yarn run build",
     "docs:watch": "cd website && yarn run start"
   },
@@ -32,50 +31,43 @@
     "react-native": "*",
     "react-native-vector-icons": "*"
   },
+  "dependencies": {
+    "@callstack/react-theme-provider": "^3.0.5",
+    "babel-plugin-module-resolver": "^3.2.0",
+    "color": "^3.1.2"
+  },
   "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/runtime": "^7.1.2",
+    "@babel/core": "^7.4.4",
+    "@babel/runtime": "^7.4.4",
+    "@callstack/eslint-config": "^9.0.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^23.6.0",
+    "babel-jest": "^24.3.1",
     "babel-plugin-flow-react-proptypes": "^6.1.0",
-    "babel-preset-react-native": "^4.0.0",
     "escape-string-regexp": "^2.0.0",
-    "eslint": "^4.19.1",
+    "eslint": "^6.7.2",
     "eslint-config-callstack-io": "^1.1.1",
     "eslint-plugin-prettier": "^3.0.0",
     "eslint-plugin-react-native": "^3.5.0",
     "flow-bin": "0.92.0",
-    "husky": "^0.14.3",
-    "jest": "^21.2.1",
-    "lint-staged": "^4.2.3",
-    "metro-react-native-babel-preset": "^0.49.0",
-    "prettier": "^1.14.3",
-    "react": "16.6.0-alpha.8af6728",
-    "react-native-vector-icons": "^4.4.2",
-    "react-test-renderer": "^16.0.0",
+    "jest": "^24.9.0",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "prettier": "^1.19.0",
+    "react": "16.9.0",
+    "react-native": "^0.61.4",
+    "react-native-vector-icons": "~6.3.0",
+    "react-test-renderer": "16.8.3",
     "svgexport": "^0.3.2"
   },
   "jest": {
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "/dist/"
-    ]
-  },
-  "dependencies": {
-    "@callstack/react-theme-provider": "^2.0.1",
-    "babel-plugin-module-resolver": "^3.2.0",
-    "color": "^2.0.1",
-    "lodash": "^4.17.4",
-    "react-native": "^0.59.4"
-  },
-  "resolutions": {
-    "**/babel-core": "7.0.0-bridge.0"
-  },
-  "lint-staged": {
-    "src/**/*.{js,jsx,json,css}": [
-      "prettier --single-quote --trailing-comma es5 --write './src/**/*.js'",
-      "git add"
+    "preset": "react-native",
+    "cacheDirectory": "./cache/jest",
+    "modulePathIgnorePatterns": [
+      "<rootDir>/example/node_modules",
+      "<rootDir>/lib/"
+    ],
+    "watchPathIgnorePatterns": [
+      "__fixtures__\\/[^/]+\\/(output|error)\\.js"
     ]
   }
 }
